requirements.txt
python-telegram-bot==20.7
opencv-python-headless==4.8.1.78
pillow==10.1.0
pytesseract==0.3.10
numpy==1.24.3
flask==2.3.3